\name{pressData-package}
\alias{pressData-package}
\docType{package}
\title{
Data sets in package PRESS--Protein REsidue-Level Structural Statistics
}
\description{
Data sets for graphically displaying distributional information about residue-level structural properties. Used to model and analyze a user provided protein structure in terms of its structural properties.
}
\details{
\tabular{ll}{
Package: \tab pressData\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2010-07-20\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
}
\author{
Yuanyuan Huang, Stephen Bonett, and Zhijun Wu

Maintainer: Yuanyuan Huang <sunnyuan.h@gmail.com>

}

\keyword{ package }

