\name{Torsion_Angle_Torsion}
\alias{Torsion_Angle_Torsion}
\docType{data}
\title{
Torsion_Angle_Torsion                                               
}
\description{
A matrix data frame refer tau(1)-alpha-tau(2) correlations.
}
\author{Yuanyuan Huang, Stephen Bonett, and Zhijun Wu}
\keyword{R data set}

